rename fetch-utils.ts to fetchUtils.ts
